feat: :sparkles: agregar validación para la eliminación de años en AnioController, asegurando que no existan centros de costo asociados antes de permitir la eliminación. También se añade la relación centroCostos en el modelo AnioCalendario.

+++ app/Http/Controllers/Payroll/AnioController.php
@@ -99,6 +99,10 @@ class AnioController extends Controller
 				return back()->withErrors(['message' => 'No se puede eliminar el año porque tiene planillas asociadas']);
 			}
 
+			if ($anio->centroCostos()->exists()) {
+				return back()->withErrors(['message' => 'No se puede eliminar el año porque tiene centros de costo asociados']);
+			}
+
 			$anio->delete();
 			return back()->with('success', 'Año eliminado correctamente');
 		} catch (\Throwable $th) {
